// src/games/remnant2/core/types.ts
type ItemModifiers = {
	accuracy: number;
	ammo: number;
	armor: number;
	armorStepPercentage: number;
	armorPercentage: number;
	crit: number;
	damage: number;
	elementalResistanceStep: number;
	health: number;
	healthCap: number;
	healthPercentage: number;
	healthStep: number;
	stamina: number;
	staminaPercentage: number;
	staminaStep: number;
	resistBleed: number;
	resistFire: number;
	resistShock: number;
	resistBlight: number;
	resistToxin: number;
	shieldStepPercentage: number;
	stagger: number;
	weakspot: number;
	weight: number;
	weightPercentage: number;
	weightThreshold: number;
	weightThresholds: number[];
};

type Remnant2LinkedItem = Partial<{
	archetype: { name: string };
	skills: Array<{ name: string }>;
	weapon: { name: string };
	mod: { name: string };
	traits: Array<{ name: string; amount: number }>;
	perks: Array<{ name: string }>;
}>;

type Remnant2InlineTags = (typeof INLINE_ITEM_TAGS)[number]["type"][];
type Remnant2CommunityTags = (typeof COMMUNITY_ITEM_TAGS)[number]["token"][];
type Remnant2SearchableTags = (typeof SEARCHABLE_ITEM_TAGS)[number][];

/**
 * Shared base item type for Remnant 2 items
 * Not intended to be used directly - use Remnant2Item or specific item types instead
 */
type BaseRemnant2Item<TItemSubcategory = string> = AppItem<
	Remnant2ItemCategory,
	TItemSubcategory,
	Remnant2InlineTags,
	Remnant2CommunityTags,
	Remnant2SearchableTags,
	Remnant2LinkedItem
> & {
	wikiUrl: string | undefined;
	dlc: Remnant2DLC;
	location: Remnant2ItemLocation | undefined;
	modifiers?: Partial<ItemModifiers>;
};
// ...
type OtherLocation =
	| "World Drop"
	| "Vendor"
	| "Quest"
	| "Aberration"
	| "Linked Item";

// Generic helper types for location definitions
type WorldDungeonLocation<W extends WorldLocation, D> = {
	world: W;
	dungeon: D[] | OtherLocation;
	biome?: never;
	injectable?: Injectable;
};

type WorldBiomeLocation<W extends WorldLocation, B> = {
	world: W;
	dungeon?: never;
	biome: B;
	injectable?: Injectable;
};

type Remnant2ItemLocation =
	| WorldDungeonLocation<"Losomn", LosomnDungeon>
	| WorldBiomeLocation<"Losomn", BiomeNameForWorld<"Losomn">>
	| WorldDungeonLocation<`N'Erud`, NErudDungeon>
	| WorldBiomeLocation<`N'Erud`, BiomeNameForWorld<`N'Erud`>>
	| WorldDungeonLocation<"Yaesha", YaeshaDungeon>
	| WorldBiomeLocation<"Yaesha", BiomeNameForWorld<"Yaesha">>
	| WorldDungeonLocation<"Root Earth", RootEarthDungeon>
	| WorldDungeonLocation<"Labyrinth", LabyrinthDungeon>
	| { world: "Ward 13"; dungeon: "Ward 13" | OtherLocation }
	| { world: "The Backrooms"; dungeon: "The Backrooms" }
	| { world: "Any"; dungeon: OtherLocation };